Fixed bug that caused algorithms to fail openMaze and improved AStar algorithm
Algorithms now check if state is duplicate upon expansion to prevent multiple of the same state being on the frontier and all of them being evaluated, which caused an infinite loop on a mze like openMaze.

AStar will now remove states from the VisitedNodes list and re-evaluate them if it has found a faster route to reach that state.